refactor(services): extract the session-adjacent web contexts
Moves 22 methods into five modules the session-state seam unblocked:
session categories, expert roles, known projects and workspaces, agent
terminals, and usage statistics. Each owns its own bindings and keeps
them unexported.

createSession still needs the project, remote-workspace and expert-role
helpers, so those are exported as behavior from the modules that own the
state, matching how split-web-agent-client handled the same shape.

web-agent-client.ts 2973 -> 2580, agent-service.ts 422 -> 392.
