+++ src/services/web-known-workspace-client.ts
@@ -1,0 +1,134 @@
+import { i18n } from "../i18n";
+import type {
+  CreateSessionInput,
+  KnownProject,
+  KnownRemoteWorkspace,
+  ProjectInspection,
+  RemoteWorkspace,
+} from "../types/agent";
+import type { KnownWorkspaceService } from "./session-organization-service";
+import { normalizeDisplayPath } from "../lib/session-path";
+import { nowIso } from "./web-mock-clock";
+
+let knownProjects: KnownProject[] = [];
+let knownRemoteWorkspaces: KnownRemoteWorkspace[] = [];
+
+function pathSegments(path: string) {
+  return path.split(/[\\/]/).filter(Boolean);
+}
+
+export function displayNameForPath(path: string) {
+  return pathSegments(path).at(-1) ?? path;
+}
+
+function parentPath(path: string) {
+  const normalized = path.replace(/[\\/]+$/, "");
+  const separatorIndex = Math.max(normalized.lastIndexOf("\\"), normalized.lastIndexOf("/"));
+  return separatorIndex <= 0 ? normalized : normalized.slice(0, separatorIndex);
+}
+
+export function joinSiblingPath(projectPath: string, worktreeName: string) {
+  const separator = projectPath.includes("\\") ? "\\" : "/";
+  return `${parentPath(projectPath)}${separator}${displayNameForPath(projectPath)}-${worktreeName}`;
+}
+
+export function validateWorktreeName(name: string) {
+  const trimmed = name.trim();
+  if (!trimmed || trimmed.includes("/") || trimmed.includes("\\") || trimmed.includes("..") || /[\u0000-\u001f]/.test(trimmed)) {
+    throw new Error("Invalid worktree name");
+  }
+  return trimmed;
+}
+
+export function inspectMockProject(path: string): ProjectInspection {
+  const trimmedPath = path.trim();
+  const isGit = !/(^|[\\/])(non-git|scratch|plain)([\\/]|$)/i.test(trimmedPath);
+  return {
+    path: trimmedPath,
+    displayName: displayNameForPath(trimmedPath),
+    isGit,
+    gitRoot: isGit ? trimmedPath : null,
+  };
+}
+
+export function upsertKnownProject(inspection: ProjectInspection) {
+  const timestamp = nowIso();
+  const project: KnownProject = {
+    path: inspection.path,
+    displayName: inspection.displayName,
+    isGit: inspection.isGit,
+    lastOpenedAt: timestamp,
+  };
+  knownProjects = [project, ...knownProjects.filter((candidate) => candidate.path !== project.path)];
+  return project;
+}
+
+export function resolveProjectPath(input: CreateSessionInput) {
+  const path = input.projectPath?.trim() || input.folder?.trim() || "";
+  return path ? normalizeDisplayPath(path) : null;
+}
+
+function displayNameForRemotePath(path: string) {
+  return path.replace(/\/+$/, "").split("/").filter(Boolean).at(-1) ?? path;
+}
+
+export function normalizeRemoteWorkspace(input: NonNullable<CreateSessionInput["remoteWorkspace"]>): RemoteWorkspace {
+  const host = input.host.trim();
+  const port = input.port ?? 22;
+  const path = input.path.trim();
+  const user = input.user?.trim() || null;
+  if (!host || !path) {
+    throw new Error("Remote workspace requires host and path");
+  }
+  if (host.includes("/") || host.includes("\\") || /[\u0000-\u001f]/.test(`${host}${path}${user ?? ""}`)) {
+    throw new Error("Invalid remote workspace");
+  }
+  if (!Number.isInteger(port) || port < 1 || port > 65535) {
+    throw new Error("Invalid remote workspace port");
+  }
+  const authority = user ? `${user}@${host}` : host;
+  const portSegment = port === 22 ? "" : `:${port}`;
+  return {
+    host,
+    port,
+    user,
+    path,
+    displayName: input.displayName?.trim() || `${host}:${displayNameForRemotePath(path)}`,
+    uri: `ssh://${authority}${portSegment}${path.startsWith("/") ? "" : "/"}${path}`,
+  };
+}
+
+export function upsertKnownRemoteWorkspace(remoteWorkspace: RemoteWorkspace) {
+  const timestamp = nowIso();
+  const known: KnownRemoteWorkspace = { ...remoteWorkspace, lastOpenedAt: timestamp };
+  knownRemoteWorkspaces = [
+    known,
+    ...knownRemoteWorkspaces.filter((candidate) => candidate.uri !== remoteWorkspace.uri),
+  ];
+  return known;
+}
+
+export const webKnownWorkspaceClient: KnownWorkspaceService = {
+  async listKnownProjects() {
+    return knownProjects.map((project) => ({ ...project }));
+  },
+
+  async listKnownRemoteWorkspaces() {
+    return knownRemoteWorkspaces.map((workspace) => ({ ...workspace }));
+  },
+
+  async inspectProject(path: string) {
+    if (!path.trim()) {
+      throw new Error(i18n.t("web.error.projectPathRequired"));
+    }
+    return inspectMockProject(path);
+  },
+
+  async selectProjectDirectory() {
+    return "D:\\\\example-workspace";
+  },
+
+  async selectWorkspaceDirectory() {
+    return "D:\\\\example-workspace";
+  },
+};
